Fixes issue #8 (caused leak when `luaWrapperNewIndex` or `setProperty:` raised an error, invoking longjmp and thereby skipping the call to objc_release.
This fix uses the __unsafe_unretained modifier to tell ARC not to call retain nor release at all in these functions, thereby circumventing the issue.

The code also now compiles properly with disabled ARC, which was mainly needed for debugging but can be left in there, just in case someone else feels the need for this.